requirements.txt
django 
djangorestframework 
djangorestframework-simplejwt 
ipykernel 
django-cors-headers 
gunicorn 
